requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.1
blinker==1.9.0
bs4==0.0.2
cbor==1.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
datasets==4.5.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
FlagEmbedding==1.3.5
FlashRank==0.2.10
Flask==3.1.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.5.0
humanfriendly==10.0
idna==3.11
ijson==3.4.0.post0
iniconfig==2.3.0
inscriptis==2.7.0
ir_datasets==0.5.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
lxml==5.4.0
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.1
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
peft==0.18.1
pillow==12.1.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdfium2==5.3.0
pyreadline3==3.5.4
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-build==0.18.1
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.2
sse-starlette==3.2.0
starlette==0.50.0
starlette-context==0.3.6
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==5.2.0
trec-car-tools==2.6
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unlzw3==0.2.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
websockets==15.0.1
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
